pyproject.toml
[tool.poetry]
name = "startrack"
version = "0.1.0"
description = ""
authors = ["“magelick” <fcdm2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.33"}
pytest = "^8.3.2"
mypy = "^1.11.2"
flake8 = "^7.1.1"
black = "^24.8.0"
asyncpg = "^0.29.0"
pre-commit = "^3.8.0"
fastapi-cache2 = "^0.2.2"
redis = "^5.0.8"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
